refactor: 클래스명 변경 및 패키지 구조 수정
